pyproject.toml
[tool.poetry]
name = "shift-ml-testtask"
version = "0.1.0"
description = ""
authors = ["RasulMakhmudov <makhmudov_ra@vk.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.9"
fastapi = "^0.97.0"
fastapi-users = {extras = ["sqlalchemy"], version = "^11.0.0"}
alembic = "^1.11.1"
psycopg2-binary = "^2.9.6"
uvicorn = "^0.22.0"
python-dotenv = "^1.0.0"
asyncpg = "^0.27.0"
sqlalchemy = "^2.0.17"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.3.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
